Dues 숫자 부분 Long으로 변경 및 그에 따른 코드 수정

+++ src/main/java/com/lshdainty/myhr/InitDB.java
@@ -517,24 +517,24 @@ public class InitDB {
         }
 
         public void initSetDues() {
-            saveDues("이서준", 10000, DuesType.BIRTH, DuesCalcType.PLUS, "20250104", "생일비");
-            saveDues("김서연", 10000, DuesType.BIRTH, DuesCalcType.PLUS, "20250104", "생일비");
-            saveDues("김지후", 10000, DuesType.BIRTH, DuesCalcType.PLUS, "20250104", "생일비");
-            saveDues("이준우", 10000, DuesType.BIRTH, DuesCalcType.PLUS, "20250104", "생일비");
-            saveDues("조민서", 80000, DuesType.BIRTH, DuesCalcType.MINUS, "20250131", "생일비 출금");
-            saveDues("이하은", 10000, DuesType.BIRTH, DuesCalcType.PLUS, "20250204", "생일비");
-            saveDues("김서연", 10000, DuesType.BIRTH, DuesCalcType.PLUS, "20250204", "생일비");
-            saveDues("김지후", 10000, DuesType.BIRTH, DuesCalcType.PLUS, "20250204", "생일비");
-            saveDues("이준우", 10000, DuesType.BIRTH, DuesCalcType.PLUS, "20250204", "생일비");
-            saveDues("조민서", 30000, DuesType.BIRTH, DuesCalcType.MINUS, "20250228", "생일비 출금");
-            saveDues("이서준", 30000, DuesType.OPERATION, DuesCalcType.PLUS, "20250128", "운영비 입금");
-            saveDues("김서연", 30000, DuesType.OPERATION, DuesCalcType.PLUS, "20250428", "운영비 입금");
-            saveDues("김지후", 10000, DuesType.FINE, DuesCalcType.PLUS, "20250728", "운영비 입금");
-            saveDues("조민서", 20000, DuesType.FINE, DuesCalcType.PLUS, "20250728", "운영비 출금");
-            saveDues("이준우", 10000, DuesType.FINE, DuesCalcType.PLUS, "20250728", "운영비 출금");
-            saveDues("이하은", 10000, DuesType.OPERATION, DuesCalcType.MINUS, "20250728", "운영비 출금");
-            saveDues("조민서", 10000, DuesType.OPERATION, DuesCalcType.MINUS, "20250728", "운영비 출금");
-            saveDues("이서준", 10000, DuesType.BIRTH, DuesCalcType.PLUS, "20250704", "생일비");
+            saveDues("이서준", 10000L, DuesType.BIRTH, DuesCalcType.PLUS, "20250104", "생일비");
+            saveDues("김서연", 10000L, DuesType.BIRTH, DuesCalcType.PLUS, "20250104", "생일비");
+            saveDues("김지후", 10000L, DuesType.BIRTH, DuesCalcType.PLUS, "20250104", "생일비");
+            saveDues("이준우", 10000L, DuesType.BIRTH, DuesCalcType.PLUS, "20250104", "생일비");
+            saveDues("조민서", 80000L, DuesType.BIRTH, DuesCalcType.MINUS, "20250131", "생일비 출금");
+            saveDues("이하은", 10000L, DuesType.BIRTH, DuesCalcType.PLUS, "20250204", "생일비");
+            saveDues("김서연", 10000L, DuesType.BIRTH, DuesCalcType.PLUS, "20250204", "생일비");
+            saveDues("김지후", 10000L, DuesType.BIRTH, DuesCalcType.PLUS, "20250204", "생일비");
+            saveDues("이준우", 10000L, DuesType.BIRTH, DuesCalcType.PLUS, "20250204", "생일비");
+            saveDues("조민서", 30000L, DuesType.BIRTH, DuesCalcType.MINUS, "20250228", "생일비 출금");
+            saveDues("이서준", 30000L, DuesType.OPERATION, DuesCalcType.PLUS, "20250128", "운영비 입금");
+            saveDues("김서연", 30000L, DuesType.OPERATION, DuesCalcType.PLUS, "20250428", "운영비 입금");
+            saveDues("김지후", 10000L, DuesType.FINE, DuesCalcType.PLUS, "20250728", "운영비 입금");
+            saveDues("조민서", 20000L, DuesType.FINE, DuesCalcType.PLUS, "20250728", "운영비 출금");
+            saveDues("이준우", 10000L, DuesType.FINE, DuesCalcType.PLUS, "20250728", "운영비 출금");
+            saveDues("이하은", 10000L, DuesType.OPERATION, DuesCalcType.MINUS, "20250728", "운영비 출금");
+            saveDues("조민서", 10000L, DuesType.OPERATION, DuesCalcType.MINUS, "20250728", "운영비 출금");
+            saveDues("이서준", 10000L, DuesType.BIRTH, DuesCalcType.PLUS, "20250704", "생일비");
         }
 
         public void saveMember(String id, String pwd, String name, String email, String birth, String workTime, String employ, String lunar) {
@@ -553,7 +553,7 @@ public class InitDB {
             em.persist(schedule);
         }
 
-        public void saveDues(String userName, int amount, DuesType type, DuesCalcType calc, String date, String detail) {
+        public void saveDues(String userName, Long amount, DuesType type, DuesCalcType calc, String date, String detail) {
             Dues dues = Dues.createDues(userName, amount, type, calc, date, detail);
             em.persist(dues);
         }
