+++ src/main/java/com/lshdainty/myhr/domain/Dues.java
@@ -18,7 +18,7 @@ public class Dues {
     private String userName;
 
     @Column(name = "dues_amount")
-    private int amount;
+    private Long amount;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "dues_type")
@@ -35,7 +35,7 @@ public class Dues {
     private String detail;
 
     // 회비 생성자  (setter말고 해당 메소드 사용할 것)
-    public static Dues createDues(String userName, int amount, DuesType type, DuesCalcType calc, String date, String detail) {
+    public static Dues createDues(String userName, Long amount, DuesType type, DuesCalcType calc, String date, String detail) {
         Dues dues = new Dues();
         dues.userName = userName;
         dues.amount = amount;
